pipelines/dask/requirements.txt
# python 3.11
dask~=2023.7.0
dask[complete]
psycopg2-binary~=2.9.0
